requirements.txt
faster-whisper
easyocr
pdf2image
google-genai
ffmpeg-python
numpy
opencv-python-headless
